+++ src/js/index.js
@@ -3,15 +3,15 @@ import "core-js/stable";
 // import "regenerator-runtime/runtime";
 import "../css/index.scss";
 import { map } from "./map";
-import { linechart } from "./linechart";
+import { createLineChart } from "./linechart";
 import { timelapse } from "./timelapse";
 import { counter } from "./counter";
 import { scroll2top } from "./back2top";
 
 document.addEventListener("DOMContentLoaded", function() {
   map();
-  linechart();
+  createLineChart();
   timelapse();
-  counter();
+  // counter();
   scroll2top();
 });
